refactor: 서버액션으로 분리
